2023 day 18: DigSite: use Math.toIntExact instead of casting directly. this has the added benefit that an exception is thrown, when there would be an overflow

+++ src/main/java/tk/vivas/adventofcode/year2023/day18/DigSite.java
@@ -175,7 +175,7 @@ class DigSite {
         int x = 0;
         int y = 0;
         for (DigInstruction digInstruction : instructionList) {
-            int amount = (int) digInstruction.amount();
+            int amount = Math.toIntExact(digInstruction.amount());
             switch (digInstruction.direction()) {
                 case UP -> {
                     y -= amount;
